Extracted feature showcase cards to separate files
