+++ dashboard/src/app/(landing)/components/featureCards/funnelsCard.tsx
@@ -1,0 +1,55 @@
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
+import { TrendingUp, TrendingDown } from 'lucide-react';
+import { useMemo } from 'react';
+
+export function FunnelsCard() {
+  const funnelSteps = useMemo(
+    () => [
+      { name: 'Read Blog Post', conversion: 100, dropOff: 60.0 },
+      { name: 'Click CTA', conversion: 40, dropOff: 30.0 },
+      { name: 'View Landing Page', conversion: 28, dropOff: 16.3 },
+      { name: 'Start Form', conversion: 15, dropOff: null },
+    ],
+    [],
+  );
+
+  return (
+    <Card className='dark:metric-card dark:shadow-card-glow'>
+      <CardHeader>
+        <CardTitle className='text-xl'>Funnels</CardTitle>
+        <CardDescription className='text-base'>
+          Track the conversion rate of your website's key user journeys.
+        </CardDescription>
+      </CardHeader>
+      <CardContent className='space-y-3'>
+        {funnelSteps.map((step, index) => (
+          <div key={index} className='space-y-1'>
+            <div className='flex items-center justify-between'>
+              <span className='text-sm font-medium'>{step.name}</span>
+              <span className='text-sm font-bold'>{step.conversion}%</span>
+            </div>
+            <div className='bg-muted h-2 w-full overflow-hidden rounded-full'>
+              <div className='bg-primary h-full' style={{ width: `${step.conversion}%` }} />
+            </div>
+            {step.dropOff && (
+              <div className='text-muted-foreground flex items-center text-xs'>
+                <TrendingDown className='mr-1 h-3 w-3' />
+                <span>{step.dropOff}% drop off</span>
+              </div>
+            )}
+          </div>
+        ))}
+
+        <div className='border-border mt-4 border-t pt-3'>
+          <div className='flex items-center justify-between text-sm'>
+            <span className='font-medium'>Total: 15%</span>
+            <div className='flex items-center text-green-500'>
+              <TrendingUp className='mr-1 h-3 w-3' />
+              <span>2.3% since last week</span>
+            </div>
+          </div>
+        </div>
+      </CardContent>
+    </Card>
+  );
+}
